+++ src/ts/app/generators/html-attribute-generator.ts
@@ -4,6 +4,12 @@
 
 export class HtmlAttributeGenerator {
 
+    /**
+     * Encode special characters ...
+     *
+     * @param value
+     */
+
     encodeValue(value: string): string {
         return value.replace(/["&<>]/g, (char) => {
             switch (char) {
@@ -21,6 +27,13 @@ export class HtmlAttributeGenerator {
         });
     }
 
+    /**
+     * generate attributes string for child objects if any
+     *
+     * @param prefix
+     * @param obj
+     */
+
     processObject(prefix: string, obj: Record<string, any>): any[] {
         return Object.entries(obj).flatMap(([key, value]) => {
             const newKey : string = prefix ? `${prefix}-${key}` : key;
@@ -34,9 +47,14 @@ export class HtmlAttributeGenerator {
      * Generate the markup ...
      *
      * @param attributes
+     * @param separator
      */
 
-    generateAttributes(attributes: Record<string, any>): string {
+    generateAttributes(attributes: Record<string, any>, separator : string = ' '): string {
+
+        if(separator === '') {
+            separator = ' ';
+        }
 
         return Object.entries(attributes)
             .flatMap(([key, value]) => {
@@ -45,7 +63,7 @@ export class HtmlAttributeGenerator {
                 } else if (Array.isArray(value)) {
                     const filteredValues = value
                         .filter(v => typeof v === 'string' || typeof v === 'number')
-                        .join(' ');
+                        .join(separator);
                     return filteredValues ? [[key, filteredValues]] : [];
                 } else if (typeof value === 'object' && value !== null) {
                     return this.processObject(key, value);
@@ -53,7 +71,7 @@ export class HtmlAttributeGenerator {
                 return [];
             })
             .map(([key, value]) => `${key}="${this.encodeValue(String(value))}"`)
-            .join(' ');
+            .join(separator);
     }